+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+* Optional field auth - [#126](https://github.com/Gravity-Core/graphism/pull/126)
+
 ## 0.8.2 (August 18th, 2022)
 
 * Optional auth - [#125](https://github.com/Gravity-Core/graphism/pull/125)